Separate rest exceptions into a new folder
