+++ server/src/main/java/com/omigost/server/rest/exception/BudgetNotFound.java
@@ -1,4 +1,4 @@
-package com.omigost.server.rest;
+package com.omigost.server.rest.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
